moved reports files to the records dir
