+++ build.gradle.kts
@@ -7,7 +7,7 @@ buildscript {
 
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
-    id("com.android.application") version "8.2.2" apply false
+    id("com.android.application") version "8.3.1" apply false
     id("org.jetbrains.kotlin.android") version "1.9.23" apply false
 }
