+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,7 +1,7 @@
 #Sun Jan 28 17:05:51 CST 2024
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.2-bin.zip
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.4-bin.zip
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
 ANDROID_HOME=C\:\\Users\\user\\AppData\\Local\\Android\\Sdk
